+++ talk/app/webrtc/mediastreaminterface.h
@@ -187,7 +187,7 @@ class AudioTrackSinkInterface {
                       int bits_per_sample,
                       int sample_rate,
                       int number_of_channels,
-                      int number_of_frames) = 0;
+                      size_t number_of_frames) = 0;
  protected:
   virtual ~AudioTrackSinkInterface() {}
 };
